+++ src/core/messenger/click.ts
@@ -1,6 +1,6 @@
 import { isString } from '@guardian/libs';
 import type { RegisterListener } from 'core/messenger';
-import { trackNativeAdLinkClick } from '../analytics/google';
+import { trackNativeAdLinkClick } from '../../analytics/google';
 
 const sendClick = (adSlot: { id: string }, linkName: string): void => {
 	trackNativeAdLinkClick(adSlot.id, linkName);
