+++ src/core/messenger/disable-refresh.ts
@@ -1,5 +1,5 @@
 import type { RegisterListener } from 'core/messenger';
-import { dfpEnv } from '../dfp/dfp-env';
+import { dfpEnv } from '../../dfp/dfp-env';
 
 // This message is intended to be used with a DFP creative wrapper.
 // For reference, the wrapper will post a message, with an iFrameId, like so:
