+++ src/core/messenger/full-width.spec.ts
