+++ src/core/messenger/full-width.ts
