+++ src/core/messenger/passback.ts
@@ -4,7 +4,7 @@ import { adSizes } from 'core/ad-sizes';
 import type { RegisterListener } from 'core/messenger';
 import { getCurrentBreakpoint } from 'detect/detect-breakpoint';
 import fastdom from 'lib/fastdom-promise';
-import { adSlotIdPrefix } from '../dfp/dfp-env-globals';
+import { adSlotIdPrefix } from '../../dfp/dfp-env-globals';
 
 type PassbackMessagePayload = { source: string };
