+++ src/core/messenger/resize.ts
