+++ src/dynamic-slots/spacefinder/spacefinder-debug-tools.ts
